+++ src/wpmMonitorStatusBar.ts
@@ -1,11 +1,15 @@
 import * as vscode from 'vscode';
 
-export class WpmMonitorStatusBar { 
+export class WpmMonitorStatusBar {
+	private interval:number = 1000;	// interval (milliseconds) in which the status bar gets updated
 	private statusBar: vscode.StatusBarItem;
-	private wpm = 0;
+	private wpm: number = 0;
+	private specialEffectThreshold: number = 80;
+	private specialEffect: string = "🔥";
 
 	constructor() {
 		this.statusBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 0);
+		this.statusBar.command = "extension.reset";	// reset on click
 		this.statusBar.show();
 
 		this.updateUI();
@@ -15,10 +19,24 @@ export class WpmMonitorStatusBar {
 		this.wpm = wpm;
 	}
 
+	public getSpecialEffect() {
+		return this.specialEffect;
+	}
+	public setSpecialEffect(specialEffect: string) {
+		this.specialEffect = specialEffect;
+	}
+
+	public getSpecialEffectThreshold() {
+		return this.specialEffectThreshold;
+	}
+	public setSpecialEffectThreshold(specialEffectThreshold: number) {
+		this.specialEffectThreshold = specialEffectThreshold;
+	}
+	
 	private updateUI() {
 		return setInterval(()=> {
-		this.statusBar.text = this.wpm.toString() + " wpm";
-		}, 1000);
+		this.statusBar.text = `${this.wpm >= this.specialEffectThreshold ? this.specialEffect + " ":""}${this.wpm.toString()} wpm`;
+		}, this.interval);
 	}
 
 	public dispose() {