+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Change Log
 
-All notable changes to the "vscode-wpm-monitor" extension will be documented in this file.
+All notable changes to the `vscode-wpm-monitor` extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## 0.2.0
 
-- Initial release
+### Added
+
+- time based wpm mode and a command prompt mode switcher
+- configuration options for length of the timeframe and number of characters to be considered for wpm calculation
+- special effect that is displayed after you reach a wpm threshold
+- option to reset the wpm display with a command or clicking the counter in the status bar
+
+## 0.1.0
+
+### Added
+
+- basic functionality and character based wpm mode